Move unit_of_work to service_layer
