+++ src/subscription/views.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 
 from subscription.serializers import SubscriptionRequestSerializer
 from subscription.service_layer.services import SubscriptionService
-from subscription.unit_of_work import DjangoUnitOfWork
+from subscription.service_layer.unit_of_work import DjangoUnitOfWork
 
 
 class SubscriptionViewSet(viewsets.ViewSet):
